+++ src/applications/vaos/new-appointment/components/ReviewPage/ReviewRequestInfo/CommunityCareSection/PreferredLanguageSection.jsx
@@ -28,7 +28,7 @@ export default function PreferredLanguageSection({ data }) {
       <div className="vads-l-row vads-u-justify-content--space-between">
         <div className="vads-u-flex--1 vads-u-padding-right--1">
           <h2 className="vads-u-font-size--h3 vads-u-margin-top--0">
-            Language you’d prefer the provider speak
+            Preferred language
           </h2>
           {
             LANGUAGES.find(language => language.id === data.preferredLanguage)
